Make it possible to build all HTML variants at the same time
... by using different suffixes:

* debian-paketmanagement.online.html
* debian-paketmanagement.allinone.html
* debian-paketmanagement.doctor.html

Build the online and allinone versions by default and also deploy
them.

Copy all needed images to the deploy directory in the deploy target.

Enable the data-uri attribute explicitly for the allinone build. It's
only enabled by default for Slidy-baed slides.

There seems to be a bug in the images macro if data-uri is set and
images are in subdirectories. Hence we override the original macros
with fixed ones.